Moved `add_run_args` to `__main__.py`
